fase 15: strict-typing pass, part 1 - misc/HA-base-class + core files (186 -> 130 errors)
Systematic mypy --strict fix pass (quality_scale.yaml's strict-typing
rule), started after confirming the root cause of the ~31 "Class cannot
subclass X (has type Any)" errors: the homeassistant package installed
in this sandbox (pinned at 2024.3.3, no newer version available on this
package index - confirmed via `pip index versions homeassistant`) ships
no py.typed marker, so mypy treats every HA class/function as untyped
regardless of its real source annotations. This is the same category of
environment ceiling as ConfigSubentryFlow and UpdateFailed's translation
kwargs found earlier this session, not something fixable from this repo.

All ~25 "Class cannot subclass X" sites (and the matching
no-untyped-call/untyped-decorator sites reachable purely through HA's
own base classes, e.g. RestoreEntity.async_added_to_hass,
homeassistant.core.callback) now carry a narrowly-scoped
`# type: ignore[misc]` / `[no-untyped-call]` / `[untyped-decorator]`
with a shared one-line reason, rather than leaving strict mode entirely
off for the whole integration.

Genuine fixes (not suppressions) in this batch:
- const.py: VENTILATION_TYPES given an explicit
  dict[str, dict[str, float | str]] annotation instead of an
  implicitly-object-valued dict, fixing a real `float * object` error.
- building_model.py: float() wrap on a `**` power expression mypy
  couldn't otherwise prove returns float.
- helpers.py: removed a redundant re-annotation of `forecast` that
  mypy treats as a second, conflicting declaration.
- optimizer.py: renamed two blocks' reused `prev_off`/`prev_sum` loop
  variables (forward DP fill, backward reconstruction, buffer
  reconstruction) - mypy's loop-back-edge narrowing was pinning the
  name's type from one block's assert and flagging the other's fresh
  unpack as an incompatible re-assignment. Real, if subtle, fix - not
  a suppression.
- entity.py: BaseUtilitySensor.native_value widened from `-> float` to
  `-> float | None` (never actually returns None itself - its own
  value always defaults through `_attr_native_value or 0.0` - but
  CoordinatorEntity subclasses genuinely do return None pre-data,
  deferring to their own `available` property, which violated Liskov
  substitution under the old, narrower declared type).
- coordinator.py: _last_price given an explicit `float | None`
  annotation instead of inferring `None` from its literal `= None`
  initializer; the thermal-v2 calibration override block now
  narrows learned_ua_w_per_k/learned_thermal_mass_kwh_per_k with
  explicit is-not-None checks alongside `calibration.applied` (which
  already guarantees this at runtime, per calibration.py's own
  `applied` property - mypy just can't see through that property
  itself) rather than assigning Optional values into float attributes.
- __init__.py: `config: dict` -> `dict[str, Any]`,
  `_trigger_first_optimization` given `-> None`, `unload_ok` explicitly
  cast to bool (async_unload_platforms returns Any in this environment).
- sensor/optimization/*.py (5 files) + total_cost_savings.py: coordinator
  params given `: Any`, native_value properties given real
  `-> float | None` signatures with explicit float()/None handling
  instead of returning Any straight from a coordinator.data.get() call.

256 tests pass, ruff/black/codespell clean. 130 errors remain (down
from 186), concentrated in config_flow.py, sensor/event_driven.py,
sensor/__init__.py and the remaining sensor/daily_utility,
sensor_thermal_* modules - continuing in the next commits.

+++ custom_components/heating_curve_optimizer/building_model.py
@@ -206,7 +206,7 @@ class EmitterConfig:
         if delta_t <= 0 or self.nominal_delta_t <= 0:
             return 0.0
         ratio = delta_t / self.nominal_delta_t
-        return self.nominal_power_kw * ratio**self.exponent
+        return float(self.nominal_power_kw * ratio**self.exponent)
 
     @classmethod
     def sized_to_building(
